hooks corrected, helper naming corrections
