+++ src/components/spreadsheet-import/index.tsx
@@ -4,8 +4,7 @@ import { ReactSpreadsheetImport } from "@wavepoint/react-spreadsheet-import";
 import { Result } from "@wavepoint/react-spreadsheet-import/types/types";
 import { StyledBackdrop } from "./index.styled";
 import { getImportFields } from "utils/import-key-mappings";
-import { useCoreModuleNavigation } from "utils/helper";
-import { useNotificationsService } from "@hooks";
+import { useCoreModuleNavigation, useNotificationsService } from "@hooks";
 
 interface csvImportPorps {
   isOpen: boolean;
