+++ src/features/orders/view/index.tsx
@@ -27,7 +27,7 @@ import { DataGrid, GridCellParams, GridColDef } from "@mui/x-data-grid";
 import { useRequestContext } from "@providers/request-provider";
 import { DataView } from "components/data-view";
 import { useRouteParams } from "typesafe-routes";
-import { getCountryByCode, getFormattedDateTime } from "utils/helper";
+import { getCountryByCode, getFormattedDateTime } from "utils/general-helper";
 import { orderFormBreadcrumbLinks } from "../constants";
 import { ModuleWrapper } from "@components/module-wrapper";
 import { GhostLink } from "@components/ghost-link";
@@ -38,7 +38,7 @@ import { useModuleWrapperContext } from "@providers/module-wrapper-provider";
 import { useFormik, FormikHelpers } from "formik";
 import zod from "zod";
 import { toFormikValidationSchema } from "zod-formik-adapter";
-import { execSubmitWithToast } from "utils/formik-helpers";
+import { execSubmitWithToast } from "utils/formik-helper";
 import { useErrorDetailsModal } from "@providers/error-details-modal-provider";
 
 export const OrderViewBase = () => {
