+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./notifications-service-hook";
+export * from "./core-module-navigation-hook";
