+++ src/utils/constants.ts
@@ -1,5 +1,5 @@
 import { rootRoute } from "lib/router";
-import { BreadcrumbLink } from "./types";
+import { BreadcrumbLink } from "types";
 
 export const dataListBreadcrumbLinks: BreadcrumbLink[] = [
   { linkText: "Dashboard", toRoute: rootRoute },
